Improve master password storage.
This commit adds support to store the secret user password and the hint inside a private file. The user password (the master password) is encrypted using a secret only the phone's secure element knows. The IV for AES is also stored.

This hint is stored in plain after the encrypted content.